+++ packages/web-components/@elements/sayt/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - SFX-191: Added the `sayt` component.
-  - This component renders `sayt` related components such as `sfx-autocomplete` and `sfx-products-sayt`.
+  - This component renders `sayt` related components such as `gbe-autocomplete` and `gbe-products-sayt`.
   - SFX-333: It emits events for requesting products and autocomplete terms.
   - SFX-248: The `@elements/events` package is used for event names and payload interfaces.
   - SFX-380: With a paired searchbox, the autocomplete selection can be changed using the Up and Down arrow keys.